Clear new product form when adding new product
This commit fixes a bug where the `quantityAddedToDish` and `quantityAddedToDishUnit` fields in the `NewProductFormViewModel` were not being reset when adding a new product. This could lead to incorrect data being used when creating a new product.

The fix is to explicitly reset these fields in the `reset` function of the view model.